==> notebooks/02_factor_engineering.ipynb <==
## Notebook 02 – Factor Engineering

This notebook creates the momentum factor using the cleaned price dataset, and exports the factor datasets for downstream analysis. This notebook's template will be used in later analyses / projects for creation of other factors.

### Step 0 - Import packages and functions

import sys, os
sys.path.append(os.path.abspath("../src"))
import numpy as np
import pandas as pd

# Helper function to compute momentum
from factor_calculations import *

### Step 1 - Load Log Returns and Price datasets

# Import prices
price = pd.read_parquet('../data/processed/price_filtered.parquet') 

### Step 2 - Compute momentum factor

# Compute momentum factor
momentum = compute_momentum(price) # Compute momentum using helper function, default lag parameter = 21 days

### Step 3 - Export factor datasets

# Save factor as parquet files
momentum.to_parquet("../data/processed/momentum.parquet")
