==> notebooks/05_walk_forward_validation.ipynb <==
import sys, os
sys.path.append(os.path.abspath("../src"))

from factor_calculations import compute_log_returns
from backtest_engine import walk_forward_ls_by_regime

import pandas as pd
import numpy as np

price = pd.read_parquet("../data/processed/price_filtered.parquet")
returns = compute_log_returns(price)

momentum = pd.read_parquet("../data/processed/momentum.parquet")
regime_vol = pd.read_parquet("../data/processed/regime_market_vol.parquet")["vol_regime"]

wf_mom_low = walk_forward_ls_by_regime(momentum, returns, regime_vol, "low_vol")
wf_mom_high = walk_forward_ls_by_regime(momentum, returns, regime_vol, "high_vol")

wf_mom_low.to_frame(name="wf_mom_low_vol").to_parquet("../data/processed/wf_mom_low.parquet")
wf_mom_high.to_frame(name="wf_mom_high_vol").to_parquet("../data/processed/wf_mom_high.parquet")
